+++ src/commands/runTestFolder.js
@@ -1,8 +1,8 @@
 const test = require('../helpers/test');
 const mix = require('../helpers/mix');
 
-function handler(folderUri) {
-  test.onTestFolder(folderUri, (folder) => mix.testPath(folder));
+function handler(context, folderUri) {
+  test.onTestFolder(context, folderUri, (folder) => mix.testPath(folder));
 }
 
 module.exports = {