+++ src/commands/watchTestAtCursor.js
@@ -1,8 +1,8 @@
 const test = require('../helpers/test');
 const mix = require('../helpers/mix');
 
-function handler() {
-  test.onTestFile((fileName, cursorLine) => mix.testFileAt(fileName, cursorLine));
+function handler(context) {
+  test.onTestFile(context, (fileName, cursorLine) => mix.testFileAt(fileName, cursorLine));
 }
 
 module.exports = {