+++ src/helpers/term.js
@@ -1,4 +1,5 @@
 const vscode = require('vscode');
+const storage = require('./storage');
 
 const config = vscode.workspace.getConfiguration('vscode-elixir-test');
 
@@ -6,10 +7,17 @@ function terminal() {
   return vscode.window.activeTerminal || vscode.window.createTerminal();
 }
 
-function run(cmd) {
+function run(context, cmd, store = true) {
   const term = terminal();
   term.sendText(cmd);
-  if (config.focusOnTerminalAfterTest) term.show();
+
+  if (config.focusOnTerminalAfterTest) {
+    term.show();
+  }
+
+  if (store) {
+    storage.update(context, 'lastCmd', cmd);
+  }
 }
 
 module.exports = {