+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - New: Added support to run test coverage 
+- New: Added suppor to run the last test command
 - New: Custom `MIX_ENV` via config file
 - Fix: `cd` the root folder to run `mix test`, `mix test.watch`...
 - Fix: Apply linter suggestions to the project